+++ pkg/host_group/repo.go
@@ -1,9 +1,9 @@
 package host_group
 
 type Repo interface {
-	Delete(id uint64) error
+	Delete(id uint32) error
 	GetAll() ([]*HostGroup, error)
-	GetByID(id uint64) (*HostGroup, error)
+	GetByID(id uint32) (*HostGroup, error)
 	Store(u *HostGroup) error
 	Update(u *HostGroup) error
 }
