+++ pkg/round/repo.go
@@ -1,9 +1,9 @@
 package round
 
 type Repo interface {
-	Delete(id uint64) error
+	Delete(id uint32) error
 	GetAll() ([]*Round, error)
-	GetByID(id uint64) (*Round, error)
+	GetByID(id uint32) (*Round, error)
 	Store(u *Round) error
 	Update(u *Round) error
 	GetLastRound() (*Round, error)
