+++ pkg/service/repo.go
@@ -1,9 +1,9 @@
 package service
 
 type Repo interface {
-	Delete(id uint64) error
+	Delete(id uint32) error
 	GetAll() ([]*Service, error)
-	GetByID(id uint64) (*Service, error)
+	GetByID(id uint32) (*Service, error)
 	Store(u *Service) error
 	Update(u *Service) error
 }
